Rework DataModel tests after main DataModel refactor.
